flow - consolidate with layout config option
