+++ src/index.js
@@ -1,3 +1,4 @@
 export { default as Visual } from './visual'
+export { default as Flow } from './flow'
 export { default as Flow2 } from './flow2'
 export { default as Flow4 } from './flow4'
